feat: a field cut into pieces after the words is highlighted on the words
Pieces cut out of whole words keep the word's offsets, so a match on a piece
marks the word it came from; only a tokenizer that cuts pieces places them
inside the words. `delimited_term_freq` reads `|` unless told otherwise,
and `type_as_payload` is known by name.

+++ src/api/doc/termvectors.rs
@@ -85,7 +85,7 @@ fn field_statistics_of(g: &IdxState, field: &str) -> (u64, u64) {
             }
         }
     }
-    (sum_doc_freq, searcher.num_docs() as u64)
+    (sum_doc_freq, searcher.num_docs())
 }
 
 pub(crate) fn term_vectors_of(
